AA shell is finished!
