+++ hypex/extensions/abstract.py
@@ -3,9 +3,9 @@ from typing import Union, Any, Dict
 
 from hypex.dataset import Dataset
 from hypex.dataset.backends import PandasDataset
+from hypex.dataset.dataset import DatasetAdapter
 from hypex.utils.errors import AbstractMethodError
 from hypex.utils import FieldKeyTypes
-from hypex.utils.adapter import Adapter
 from hypex.dataset import ABCRole
 
 
@@ -26,7 +26,7 @@ class Extension(ABC):
     def result_to_dataset(
         result: Any, roles: Union[ABCRole, Dict[FieldKeyTypes, ABCRole]]
     ) -> Dataset:
-        return Adapter.to_dataset(result, roles=roles)
+        return DatasetAdapter.to_dataset(result, roles=roles)
 
 
 class CompareExtension(Extension, ABC):
