+++ hypex/utils/enums.py
@@ -34,3 +34,10 @@ class ABNTestMethodsEnum(enum.Enum):
     fdr_tsbh = "fdr_tsbh"
     fdr_tsbky = "fdr_tsbky"
     quantile = "quantile"
+
+
+@enum.unique
+class RenameEnum(enum.Enum):
+    all = "all"
+    columns = "columns"
+    index = "index"
